корректировка классов заказа

+++ src/main/java/com/store/services/OrderService.java
@@ -2,14 +2,24 @@ package com.store.services;
 
 import com.api.dto.OrderDTO;
 import com.api.mappers.OrderMapper;
+import com.store.entities.Status;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.persistence.criteria.Predicate;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.store.entities.OrderEntity;
 import com.store.repositories.OrderRepository;
 
+import java.time.Instant;
 import java.util.ArrayList;
+import java.util.EnumSet;
 import java.util.List;
+import java.util.Set;
 
 @Service
 @Transactional
@@ -18,28 +28,63 @@ public class OrderService {
 
     private final OrderRepository orderRepository;
     private final OrderMapper orderMapper;
+    private final Set<Status> statusSet = EnumSet.allOf(Status.class);
 
     public OrderDTO createOrder(OrderDTO order) {
+         if (order.getStatus() == null) {
+             throw new IllegalArgumentException("Статус обязателен");
+         }
+         Status status = order.getStatus();
+         if (!statusSet.contains(status)) {
+             throw new IllegalArgumentException("Статус должен быть: " + Status.listAll());
+         }
+
         OrderEntity orderEntity = orderMapper.toOrderEntity(order);
-        OrderEntity savedOrderEntity = orderRepository.save(orderEntity);
-        return orderMapper.toOrderDTO(savedOrderEntity);
+        return orderMapper.toOrderDTO(orderRepository.save(orderEntity));
     }
 
-    public OrderDTO updateOrder(OrderDTO order) {
-        OrderEntity orderEntity = orderMapper.toOrderEntity(order);
-        OrderEntity savedOrderEntity = orderRepository.save(orderEntity);
-        return orderMapper.toOrderDTO(savedOrderEntity);
+    public OrderDTO updateOrder(int id, Status status) {
+
+        OrderEntity orderEntity = orderRepository
+                .findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Заказ не найден"));
+
+        if (status == null) {
+            throw new IllegalArgumentException("Статус обязателен");
+        }
+        if (!statusSet.contains(status)) {
+            throw new IllegalArgumentException("Статус должен быть: " + Status.listAll());
+        }
+
+        orderEntity.setStatus(status);
+        return orderMapper.toOrderDTO(orderRepository.save(orderEntity));
     }
 
-    public List<OrderDTO> getAllOrders() {
-        List<OrderDTO> orderDTOS = new ArrayList<>();
-        orderRepository.findAll().forEach(orderEntity -> {
-            orderDTOS.add(orderMapper.toOrderDTO(orderEntity));
+    public Page<OrderDTO> getAllOrders(Instant createdAt, Status status, Pageable pageable) {
+
+        Specification<OrderEntity> spec = ((root, query, criteriaBuilder) -> {
+            List<Predicate> predicates = new ArrayList<>();
+
+            if (createdAt != null) {
+                predicates.add(criteriaBuilder.equal(root.get("createdAt"), createdAt));
+            }
+            if (status != null) {
+                predicates.add(criteriaBuilder.equal(root.get("status"), status));
+            }
+
+            return predicates.isEmpty()
+                    ? null
+                    : criteriaBuilder.and(predicates.toArray(new Predicate[0]));
         });
-        return orderDTOS;
+        return orderRepository
+                .findAll(spec, pageable)
+                .map(orderMapper::toOrderDTO);
     }
 
     public void deleteOrder(int id) {
+        if (!orderRepository.existsById(id)) {
+            throw new EmptyResultDataAccessException("Нет заказа с id: " + id, 1);
+        }
         orderRepository.deleteById(id);
     }
 }
